Second set of updates for recent discussions, TAG comments, etc
Changes:
-	Added "7.4 Push Registration Uniqueness" to clarify for TAG comment
that same-origin webpages do not share registrations.
-	Added text to "7.3 Push Registration Persistence" to clarify for TAG
comment that how user-agents invoke inactive webapp is not described
here... expected in Service Workers
-	Added text to "3. Terminology" and "4. Security and privacy
considerations" for TAG comment, to clarify what is meant by "express
permission"
- In "7.2 Steps", per TAG comment removed the text "or alternatively
that the requesting webapp has permission to deactivate such push
registration", focusing the NoModificationAllowedError purely on the
validity of the passed pushRegistrationId for this webapp.
-	Added notes for consideration of various TAG comments and suggestions.
- Added text to "7.4 Push Registration Uniqueness" for TAG comment, to
clarify that webapp association of specific registrations to "message
streams or sources" is an application-layer concern, and that webapp
servers can provide means to map registrations to specific webapp
features/functions as needed.
-	Fixed various missing anchors to code elements
-	Fixed respec issues
-	Fixed normative vs informative reference issues